Finish SignupForm layout
I added a loginpage with a single button called Create Account.
when the button is clicked, a dialog appears with the signup form.
Since the API doesn't exist yet, the create button does nothing,
but the cancel button closes the dialog.

i also changed the whole app to a dark theme, but we should probably
come up with a better theme of our own later